Fix sign issue
Seems the original was "The signs are most already done." or something like this.

"Signs" means "sinais" only if it was the meaning of "signals".
I never heard someone calling what the character was holding a "sinal".
To mean "signboard" the word is "placa" or "letreiro".
I translated to "cartazes", which means "poster", because it makes sense in the context as you can understand "the poster is in a signboard".
Also because, as I don't think he said that sentence then it's an adaption. If I'm free to adapt then I use the word I wish, if makes sense, of course.